fix(notifications): On agent is_online status user does not notified
This caused because of the changes on the consumer. The signals weren't working because the pre_save signal attempted to query the database while the row was locked by select_for_update() within the transaction.

We are fixing that by moving the notification logic from the signals to the consumer